requirements.txt
# Core data science
numpy>=1.25
pandas>=2.1
scikit-learn>=1.3

# Chemoinformatics
rdkit-pypi>=2023.9

# Visualization
matplotlib>=3.8
seaborn>=0.12
plotly>=5.15

# Utilities
tqdm>=4.66
pathlib>=1.0.1

# Optional: for Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.30

# To install all dependencies, use:
pip install -r requirements.txt